+++ src/layman/common/util.py
@@ -1,0 +1,9 @@
+from collections import defaultdict
+
+
+def merge_infos(infos):
+    result_infos = defaultdict(dict)
+    for source in infos:
+        for (name, info) in source.items():
+            result_infos[name].update(info)
+    return result_infos