+++ src/layman/layer/filesystem/input_chunk.py
@@ -35,7 +35,7 @@ def delete_layer(username, layername):
 
 get_layer_info = input_file.get_layer_info
 
-get_layer_names = input_file.get_layer_names
+get_layer_infos = input_file.get_layer_infos
 
 update_layer = input_file.update_layer