+++ src/layman/layer/geoserver/sld.py
@@ -11,6 +11,8 @@ from flask import g, current_app
 from layman.layer.filesystem.input_sld import get_layer_file
 from layman.http import LaymanError
 from layman import settings, patch_mode
+from layman.common import util as layman_util
+from . import headers_json
 from . import headers_json, headers_xml
 from . import wms
 from ...util import url_for
@@ -69,15 +71,16 @@ def get_layer_info(username, layername):
     return info
 
 
-def get_layer_names(username):
-    return []
+def get_layer_infos(username):
+    return {}
 
 
 def get_publication_names(username, publication_type):
     if publication_type != '.'.join(__name__.split('.')[:-2]):
-        raise Exception(f'Unknown pyblication type {publication_type}')
+        raise Exception(f'Unknown publication type {publication_type}')
 
-    return []
+    infos = get_layer_infos(username)
+    return list(infos)
 
 
 def get_publication_uuid(username, publication_type, publication_name):