+++ src/layman/layer/geoserver/wfs.py
@@ -13,6 +13,7 @@ from layman import settings, patch_mode
 from layman.cache import mem_redis
 from layman.layer.filesystem import input_file
 from layman.layer.util import is_layer_task_ready
+from layman.common import util as layman_util
 
 FLASK_PROXY_KEY = f'{__name__}:PROXY:{{username}}'
 
@@ -113,10 +114,10 @@ def get_wfs_proxy(username):
         return wfs_proxy
 
     def currently_changing():
-        layernames = input_file.get_layer_names(username)
+        layerinfos = input_file.get_layer_infos(username)
         result = any((
             not is_layer_task_ready(username, layername)
-            for layername in layernames
+            for layername in layerinfos
         ))
         return result
 
@@ -152,15 +153,13 @@ def get_layer_info(username, layername):
     }
 
 
-def get_layer_names(username):
+def get_layer_infos(username):
     wfs = get_wfs_proxy(username)
     if wfs is None:
-        result = []
+        result = {}
     else:
-        result = [
-            wfs_layername.split(':')[1]
-            for wfs_layername in [*wfs.contents]
-        ]
+        result = {name.split(':')[1]: {"name": name.split(':')[1],
+                                       "title": info.title} for (name, info) in wfs.contents.items()}
     return result
 
 
@@ -168,7 +167,8 @@ def get_publication_names(username, publication_type):
     if publication_type != '.'.join(__name__.split('.')[:-2]):
         raise Exception(f'Unknown pyblication type {publication_type}')
 
-    return get_layer_names(username)
+    infos = get_layer_infos(username)
+    return list(infos)
 
 
 def get_publication_uuid(username, publication_type, publication_name):