+++ test/process_client.py
@@ -1,6 +1,6 @@
+import time
 import requests
 import os
-import time
 
 from layman import settings